// generp/frontend/types/erp.ts
export interface FormConfig {
  type: "form";
  title: string;
  json_schema: Record<string, unknown>;
  ui_schema: Record<string, unknown>;
}